Small Tower adjustments
Changed cost constructor param type to double to
reflect the type of the field. Added a clarification
regarding units to the docs for the reload speed.

+++ src/main/java/seng201/team56/models/Tower.java
@@ -18,10 +18,10 @@ public class Tower implements Purchasable{
      * Sets the default level to 0
      * @param type the ResourceType of the new Tower
      * @param amount the amount the tower fills each reload
-     * @param speed the reload speed of the tower
+     * @param speed the reload speed of the tower (time between reloads in seconds)
      * @param cost the tower's cost
      */
-    public Tower(ResourceType type, int amount, float speed, int cost) {
+    public Tower(ResourceType type, int amount, float speed, double cost) {
         this.resourceType = type;
         this.resourceFullAmount = amount;
         this.reloadSpeed = speed;
